Refactor model file to separate file
