+++ podcasts/New Search/Views/Results/SearchResultCell.swift
@@ -2,70 +2,6 @@ import SwiftUI
 import PocketCastsServer
 import PocketCastsDataModel
 import PocketCastsUtils
-import SwiftUI
-import Combine
-
-class SearchResultCellModel: ObservableObject, MainEpisodeActionViewDelegate {
-
-    var episode: EpisodeSearchResult?
-    var podcastFolder: PodcastFolderSearchResult?
-    private(set) var realEpisode: BaseEpisode?
-
-    @Published var refreshTrigger: Bool = true
-
-    init(episode: EpisodeSearchResult?, podcastFolder: PodcastFolderSearchResult?) {
-        self.episode = episode
-        self.podcastFolder = podcastFolder
-        setupObservers()
-    }
-
-    func playTapped() {
-        guard let episode else {
-            return
-        }
-        PlaybackManager.shared.playEpisodeSearchResult(episode)
-    }
-
-    func pauseTapped() {
-        PlaybackActionHelper.pause()
-    }
-
-    func downloadTapped() {}
-
-    func stopDownloadTapped() {}
-
-    func errorTapped() {}
-
-    func waitingForWifiTapped() {}
-
-    private var cancellables = Set<AnyCancellable>()
-
-    private func setupObservers() {
-        Publishers.Merge3(
-            NotificationCenter.default.publisher(for: Constants.Notifications.playbackStarted),
-            NotificationCenter.default.publisher(for: Constants.Notifications.playbackEnded),
-            NotificationCenter.default.publisher(for: Constants.Notifications.playbackPaused),
-        )
-        .receive(on: OperationQueue.main)
-        .sink(receiveValue: { [unowned self] _ in
-            self.refreshTrigger.toggle()
-        })
-        .store(in: &cancellables)
-
-        NotificationCenter.default.publisher(for: Constants.Notifications.playbackPositionSaved)
-        .receive(on: OperationQueue.main)
-        .sink(receiveValue: { [unowned self] notification in
-            guard let episodeUUID = notification.object as? String,
-                  episodeUUID == episode?.uuid,
-                  let realEpisode = DataManager.sharedManager.findBaseEpisode(uuid: episodeUUID) else {
-                return
-            }
-            self.realEpisode = realEpisode
-            self.refreshTrigger.toggle()
-        })
-        .store(in: &cancellables)
-    }
-}
 
 struct SearchResultCell: View {
     @EnvironmentObject var theme: Theme
